mcp-server-template/pyproject.toml
[tool.poetry]
name = "mcp-server-template"
version = "1.0.0"
description = "A robust, modular template for creating Python-based MCP servers using FastMCP."
authors = ["Senior Software Architect AI"]
readme = "README.md"
packages = [{include = "mcp_server_template", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
mcp = {extras = ["cli"], version = "^1.2.0"}
pydantic-settings = "^2.3.4"
colorlog = "^6.8.2"
pillow = "^11.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"